moved texture creation and uniform value setting from Application to ForwardRenderer
