+++ src/scene/forward-renderer.js
@@ -6,6 +6,7 @@ import { Mat3 } from '../math/mat3.js';
 import { Mat4 } from '../math/mat4.js';
 import { Quat } from '../math/quat.js';
 import { Vec3 } from '../math/vec3.js';
+import { math } from '../math/math.js';
 
 import { BoundingBox } from '../shape/bounding-box.js';
 
@@ -20,7 +21,8 @@ import {
     PRIMITIVE_TRIANGLES,
     SEMANTIC_ATTR, SEMANTIC_POSITION,
     STENCILOP_KEEP,
-    TEXHINT_SHADOWMAP
+    TEXHINT_SHADOWMAP,
+    TEXTURETYPE_DEFAULT
 } from '../graphics/graphics.js';
 import { createShaderFromCode } from '../graphics/program-lib/utils.js';
 import { drawQuadWithShader } from '../graphics/simple-post-effect.js';
@@ -511,6 +513,9 @@ function ForwardRenderer(graphicsDevice) {
     this.ambientColor = new Float32Array(3);
 
     this.cameraParams = new Float32Array(4);
+
+    // placeholder texture for area light LUTs
+    this._createAreaLightPlaceholderLuts();
 }
 
 function mat3FromMat4(m3, m4) {
@@ -836,6 +841,135 @@ Object.assign(ForwardRenderer.prototype, {
                        (camera._clearStencilBuffer ? CLEARFLAG_STENCIL : 0),
                 stencil: camera._clearStencil
             }); // clear full RT
+        }
+    },
+
+    // placeholder LUT textures for area light
+    _createAreaLightPlaceholderLuts: function () {
+        var placeholderLutTex =  new Texture(this.device, {
+            width: 2,
+            height: 2,
+            format: PIXELFORMAT_R8_G8_B8_A8
+        });
+        placeholderLutTex.name = 'placeholder';
+
+        var pixels = placeholderLutTex.lock();
+        pixels.fill(0);
+        placeholderLutTex.unlock();
+
+        this.device.scope.resolve('areaLightsLutTex1').setValue(placeholderLutTex);
+        this.device.scope.resolve('areaLightsLutTex2').setValue(placeholderLutTex);
+    },
+
+    // creates LUT texture used by area lights
+    _uploadAreaLightLuts: function (resource) {
+        this._areaLightLuts = {
+            data1: new Float32Array(resource, 0, 16384),
+            data2: new Float32Array(resource, 16384 * 4, 16384),
+            version: new Float32Array(resource, 32768 * 4, 1)[0]
+        };
+
+        function createTexture(device, data, format) {
+            var tex = new Texture(device, {
+                width: 64,
+                height: 64,
+                format: format,
+                addressU: ADDRESS_CLAMP_TO_EDGE,
+                addressV: ADDRESS_CLAMP_TO_EDGE,
+                type: TEXTURETYPE_DEFAULT,
+                magFilter: FILTER_LINEAR,
+                minFilter: FILTER_NEAREST,
+                anisotropy: 1
+            });
+
+            tex.lock().set(data);
+            tex.unlock();
+            tex.upload();
+
+            return tex;
+        }
+
+        function offsetScale(data, offset, scale) {
+
+            var count = data.length;
+            var ret = new Float32Array(count);
+            for (var i = 0; i < count; i++) {
+                var n = i % 4;
+                ret[i] = (data[i] + offset[n]) * scale[n];
+            }
+            return ret;
+        }
+
+        function convertToHalfFloat(data) {
+
+            var count = data.length;
+            var ret = new Uint16Array(count);
+            var float2Half = math.float2Half;
+            for (var i = 0; i < count; i++) {
+                ret[i] = float2Half(data[i]);
+            }
+
+            return ret;
+        }
+
+        function convertToUint(data){
+
+            var count = data.length;
+            var ret = new Uint8ClampedArray(count);
+            for (var i = 0; i < count; i++) {
+                ret[i] = data[i] * 255;
+            }
+
+            return ret;
+        }
+
+        // create texture if not already
+        var luts = this._areaLightLuts;
+        if (luts && !luts.ready) {
+            if (luts.version !== 0) {
+                console.warn(`areaLightData version: ${luts.version} is not yet supported!`);
+            }
+
+            luts.ready = true;
+            var device = this.device;
+            var data1, data2;
+            var format = device._areaLightLutFormat;
+
+            // pick format for lut texture
+            if (format === PIXELFORMAT_RGBA32F) {
+
+                // float
+                data1 = luts.data1;
+                data2 = luts.data2;
+
+            } else if (format === PIXELFORMAT_RGBA16F) {
+
+                // half float
+                data1 = convertToHalfFloat(luts.data1);
+                data2 = convertToHalfFloat(luts.data2);
+
+            } else {
+
+                // low precision format
+                // offset and scale to avoid clipping and increase precision - this is undone in the shader
+
+                var o1 = [0.0, 0.2976, 0.01381, 0.0];
+                var s1 = [0.999, 3.08737, 1.6546, 0.603249];
+
+                var o2 = [-0.306897, 0.0, 0.0, 0.0];
+                var s2 = [1.442787, 1.0, 1.0, 1.0];
+
+                data1 = convertToUint(offsetScale(luts.data1, o1, s1));
+                data2 = convertToUint(offsetScale(luts.data2, o2, s2));
+
+            }
+
+            var tex1 = createTexture(device, data1, format);
+            var tex2 = createTexture(device, data2, format);
+
+            // assign to scope variables
+            device.scope.resolve('areaLightsLutTex1').setValue(tex1);
+            device.scope.resolve('areaLightsLutTex2').setValue(tex2);
         }
     },
