+++ src/scene/light.js
@@ -7,12 +7,6 @@ import { Vec3 } from '../math/vec3.js';
 import { Vec4 } from '../math/vec4.js';
 
 import {
-    PIXELFORMAT_RGBA32F, PIXELFORMAT_RGBA16F,
-    TEXTURETYPE_DEFAULT, ADDRESS_CLAMP_TO_EDGE, FILTER_LINEAR, FILTER_NEAREST
-} from '../graphics/graphics.js';
-import { Texture } from '../graphics/texture.js';
-
-import {
     BLUR_GAUSSIAN,
     LIGHTTYPE_DIRECTIONAL, LIGHTTYPE_OMNI, LIGHTTYPE_SPOT,
     MASK_LIGHTMAP,
@@ -327,8 +321,6 @@ class Light {
 
         this.key = key;
     }
-
-
 
     get type() {
         return this._type;
